refactor(gui): extract wiring tests to sibling file
`primer-gui/src/wiring.rs` was 1015 lines with a ~423-line
`#[cfg(test)] mod tests` block (lines 590–1015) over the GUI session
wiring helpers. Move that module verbatim into a new sibling
`src/wiring/tests.rs`.

`wiring` is a non-root module (`pub mod wiring;` in lib.rs), so its
submodule resolves to `src/wiring/tests.rs`.

wiring.rs drops 1015 → 591 lines (−424). The single parent change is the
`mod tests { … }` block → `mod tests;` declaration. As a descendant
module the tests still reach every private parent item via `use super::*`
— zero visibility churn, no signature changes.

Pure test-only relocation: no behaviour change, no incidental fixes.
gui lib tests: 205 (unchanged); workspace: 51 ok blocks / 0 failures;
clippy + fmt clean.
